Adobted Python data type lengths
both integers and floating point numbers are provided as 64bit

+++ OpenCl_DNN/src/wrapper/NeuralNetwork.h
@@ -2,24 +2,26 @@
 #include <cstring>
 #include <cstdint>
 #include <fstream>
+//#include "CLMatrix.hpp"
 
 class NeuralNetwork {
-	int64_t* result = 0;
-	int64_t layerCount;
-	int64_t* layerSize = 0;
-	int64_t* actFunctions = 0;
+	double* result = 0;
+	uint64_t layerCount;
+	uint64_t* layerSize = 0;
+	uint64_t* actFunctions = 0;
 	float learningRate;
 	float momentum;
+	//FeedForwardNN* network = 0;
 
 public:
-	NeuralNetwork(int64_t layerCount, int64_t* layerSize, int64_t* actFunctions, float learningRate, float momentum);
+	NeuralNetwork(uint64_t layerCount, uint64_t* layerSize, uint64_t* actFunctions, float learningRate, float momentum);
 	NeuralNetwork(std::string saveFile);
 	~NeuralNetwork();
-	int64_t save(std::string saveFile);
-	int64_t getResultNode(int64_t node);
-	void calc(int64_t* inputValues);
-	void train(int64_t* inputValues, int64_t* outputValues);
-	int64_t getOutputSize();
-	int64_t getInputSize();
+	uint64_t save(std::string saveFile);
+	double getResultNode(uint64_t node);
+	void calc(double* inputValues);
+	void train(double* inputValues, double* outputValues);
+	uint64_t getOutputSize();
+	uint64_t getInputSize();
 
 };
